services/agent-runtime/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-pubsub==2.18.0
google-cloud-secret-manager==2.16.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-aiplatform==1.36.0
google-cloud-run==0.10.0
google-cloud-monitoring==2.16.0
google-cloud-trace==1.11.0
google-cloud-tasks==2.14.0
google-generativeai==0.3.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-gcp-trace==1.6.0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-grpc==0.42b0
httpx==0.25.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
structlog==23.2.0
aiofiles==23.2.1
requests==2.31.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
redis==5.0.1
celery==5.3.4
docker==6.1.3
pyyaml==6.0.1
python-json-logger==2.0.7